requirements.txt
discord-py-interactions
sqlalchemy
solana
solders